# back/swagger.yaml
openapi: 3.0.0
info:
  title: AutoBid Car Auction Platform API
  description: |
    Complete API documentation for the AutoBid car auction platform.
    This API supports deposits, proxy bids, live auctions, seller approvals, payments, pickups, and title transfers.
  version: 1.0.0
  contact:
    name: AutoBid API Support
    email: khelifisalmen9@gmail.com
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.autobid.com
    description: Production server
...
paths:
...
  /bid/placeBid:
    post:
      tags: [Bidding]
      summary: Place a bid on an item
      description: Place a bid on an auction item. First bid must be >= opening bid.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - itemId
                - bidAmount
              properties:
                userId:
                  type: integer
                itemId:
                  type: integer
                bidAmount:
                  type: number
                  format: decimal
      responses:
        '200':
          description: Bid placed successfully
        '400':
          description: Bad request (insufficient bidding power, invalid amount, etc.)
...